+++ cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Introduced `launch` command for just launching Chrome as automation
+  would
 - Added `--user-agent` option for configuring the user agent to use
 - Added `--remove-selector` option for removing elements before a
   screenshot is taken